requirements.txt
langchain
openai
sentence-transformers
faiss-cpu
streamlit
python-dotenv
